move types to commont folder
