+++ x-pack/plugins/ml/public/application/data_frame_analytics/common/analytics.ts
@@ -12,6 +12,11 @@ import { cloneDeep } from 'lodash';
 import { ml } from '../../services/ml_api_service';
 import { Dictionary } from '../../../../common/types/common';
 import { extractErrorMessage } from '../../../../common/util/errors';
+import {
+  ClassificationEvaluateResponse,
+  EvaluateMetrics,
+  TrackTotalHitsSearchResponse,
+} from '../../../../common/types/data_frame_analytics';
 import { SavedSearchQuery } from '../../contexts/ml';
 import {
   AnalysisConfig,
@@ -131,49 +136,6 @@ export interface RegressionEvaluateResponse {
   };
 }
 
-export interface PredictedClass {
-  predicted_class: string;
-  count: number;
-}
-
-export interface ConfusionMatrix {
-  actual_class: string;
-  actual_class_doc_count: number;
-  predicted_classes: PredictedClass[];
-  other_predicted_class_doc_count: number;
-}
-
-export interface RocCurveItem {
-  fpr: number;
-  threshold: number;
-  tpr: number;
-}
-
-interface EvalClass {
-  class_name: string;
-  value: number;
-}
-
-export interface ClassificationEvaluateResponse {
-  classification: {
-    multiclass_confusion_matrix?: {
-      confusion_matrix: ConfusionMatrix[];
-    };
-    recall?: {
-      classes: EvalClass[];
-      avg_recall: number;
-    };
-    accuracy?: {
-      classes: EvalClass[];
-      overall_accuracy: number;
-    };
-    auc_roc?: {
-      curve?: RocCurveItem[];
-      value: number;
-    };
-  };
-}
-
 interface LoadEvaluateResult {
   success: boolean;
   eval: RegressionEvaluateResponse | ClassificationEvaluateResponse | null;
@@ -261,13 +223,6 @@ export const isClassificationEvaluateResponse = (
       arg?.classification?.auc_roc !== undefined)
   );
 };
-
-export interface UpdateDataFrameAnalyticsConfig {
-  allow_lazy_start?: string;
-  description?: string;
-  model_memory_limit?: string;
-  max_num_threads?: number;
-}
 
 export enum REFRESH_ANALYTICS_LIST_STATE {
   ERROR = 'error',
@@ -434,21 +389,6 @@ export enum REGRESSION_STATS {
   HUBER = 'huber',
 }
 
-interface EvaluateMetrics {
-  classification: {
-    accuracy?: object;
-    recall?: object;
-    multiclass_confusion_matrix?: object;
-    auc_roc?: { include_curve: boolean; class_name: string };
-  };
-  regression: {
-    r_squared: object;
-    mse: object;
-    msle: object;
-    huber: object;
-  };
-}
-
 interface LoadEvalDataConfig {
   isTraining?: boolean;
   index: string;
@@ -530,16 +470,6 @@ export const loadEvalData = async ({
     return results;
   }
 };
-
-interface TrackTotalHitsSearchResponse {
-  hits: {
-    total: {
-      value: number;
-      relation: string;
-    };
-    hits: any[];
-  };
-}
 
 interface LoadDocsCountConfig {
   ignoreDefaultQuery?: boolean;
