+++ x-pack/plugins/ml/public/application/data_frame_analytics/common/index.ts
@@ -5,13 +5,7 @@
  * 2.0.
  */
 
-export type {
-  UpdateDataFrameAnalyticsConfig,
-  IndexPattern,
-  RegressionEvaluateResponse,
-  Eval,
-  SearchQuery,
-} from './analytics';
+export type { IndexPattern, RegressionEvaluateResponse, Eval, SearchQuery } from './analytics';
 export {
   getAnalysisType,
   getDependentVar,
