+++ x-pack/plugins/ml/public/application/data_frame_analytics/pages/analytics_exploration/components/classification_exploration/get_roc_curve_chart_vega_lite_spec.tsx
@@ -16,7 +16,7 @@ import { i18n } from '@kbn/i18n';
 
 import { LEGEND_TYPES } from '../../../../../components/vega_chart/common';
 
-import { RocCurveItem } from '../../../../common/analytics';
+import { RocCurveItem } from '../../../../../../../common/types/data_frame_analytics';
 
 const GRAY = euiPaletteGray(1)[0];
 const BASELINE = 'baseline';
