+++ x-pack/plugins/ml/public/application/data_frame_analytics/pages/analytics_exploration/components/classification_exploration/use_roc_curve.ts
@@ -10,10 +10,10 @@ import { useState, useEffect } from 'react';
 import {
   isClassificationEvaluateResponse,
   ResultsSearchQuery,
-  RocCurveItem,
   ANALYSIS_CONFIG_TYPE,
 } from '../../../../common/analytics';
 import { isKeywordAndTextType } from '../../../../common/fields';
+import { RocCurveItem } from '../../../../../../../common/types/data_frame_analytics';
 
 import {
   getDependentVar,
